Implement `Board` and a semi playable framework
Hard dropping is currently broken, and there is no mechanism for DAS/ARR
of any kind, this should likely be factored into a `Game` or `InputState`
object such that state can be kept track of easier. There also needs to
be a mechanism for locking pieces after a set amount of ticks, either by
keeping track of them in game to count the number of ticks a piece is in
contact with the ground or another piece.

There is also currently no user input nor any system to handle user
input, this could be done with stdin for a mvp but that would not allow
for any keyholding.